Add a first_map_only parameter so we keep reusing the first received tatic map.
We need this for a rather exotic case: when using https://github.com/robotics-in-concert/rocon_multimaster/tree/hydro-devel/rocon_gateway, the first time we subscribe we get the map, but not in subsequent subscriptions due to clearing costmaps.
An easier solution is just to call shutdown in the map subscriber, but we still have a big delay until receiving the map, so this flag is much better option.
I didn't verify that it doesn't interfere with map updates (because is the first time I see this feature). Appart from this, we have the change running happily in our robots.