Fixes an issue with having multiple static layers
If you have a static layer in both the local and global costmaps that
use the same map topic, there is a race condition that can cause the
static layer to get stuck after printing `Requesting map....`. This race
condition seems to be due to the call to shutdown in deactivate and how
the NodeHandle handles multiple subscribers under the hood.

This issue appears to happen about 1 in 1000 times in the setup I was
testing. This fix has never failed in over 1000000 tests. Instead of
calling activate and deactivate, the publisher is only recreated if the
topic has changed. Otherwise, it reuses the old setup.