Added window size parameter to orientation filter
Previously, the orientation was calculated using the current and the
next point. However, when the path is somewhat jumpy, this results in
poor orientations. By adding this parameter and altering the orientation
calculation, the calculated orientation can be smoothened along the path
by taking into account a larger window. The orientation of index point i
will be calculated using the positions of i - window_size and i +
window_size.