costmap_2d: add virtual, new APIs, costmap lockable
Add virtual to methods to allow the Costmap2DROS and Costmap2D
objects to be sub-classed and functionality overridden.

Add APIs to query the cost and clear axis-aligned bounding-boxes to
prevent the need for direct access of the costmap bitmap.

Also make the Costmap2DROS follow lockable so it can be used in
standard RAII lock types.