costmap_3d: add third "milli" cache in queries
Add a third in-between cache called the "milli" cache.
This cache is looked up after the micro-cache.
If the resulting new distance is under a threshold, use it as a fast
path like the micro cache. Otherwise, use it as a very good bounds
for the distance search in FCL. The milli-cache binning parameters
can be looser than the micro-cache as the error is limited by the
milli-distance threshold.

This greatly improves performance of planners that generally keep the
queried poses away from objects by allowing a coarser-grained
fast-path cache for such cases, and improves misses by giving a
better upper-bound to start the full distance search.