costmap_3d: add thread-safe version of crop_hull
Only add the functionality we need, which is a 3D point inside mesh
test. Make it thread safe by not storing the input in the object.
Create one crop hull object inside every query to use for this test.
This saves much time copying mesh triangle information on every query.