costmap_3d: match size only if 3D costmap valid
Move the size match and costmap update increment after checking if we
have a 3D costmap in updateMap. The costmap only does not exist on
startup until we have been dynamically reconfigured the first time.
The costmap 2D code launches the map update thread prior to us
getting our first dynamic reconfigure (since it gets reconfigured
first). This makes a race where we can update the map before we have
a valid map. Losing this race means we setup an incorrecty sized
costmap in the z-dimension (since our dynamic reconfigure has not yet
run). This causes us to run out of memory doing the deleteAABB on the
universal bounds map when using a tree with 32 bit keys.

Fix this by only attempting to match if we have a valid map.