costmap_3d: fix interior collisions in queries
When there are two octomap cells, one near the mesh on the outside,
and one deeper in the inside, the fcl::distance algorithm (which
treats the mesh as a surface) returns the outside box. This causes
us to mis-report the collision (and signed distance) as positive.

Fix this by implementing our own octree/mesh-volume solver. The
solver relies on an oracle function (implemented using a LUT) to
know if a box is interior to the mesh. Also, to correctly handle
signed distance, the solver takes an optional signed distance
function.

Construct the LUT anew whenever the mesh is changed, but otherwise
cache it.

Also, to speed up the necessary mesh/halfspace distance checks,
cache the mesh halfspaces when the model is created.