costmap_3d: bin query orientation using cayley transform
To lookup entries in the caches, we must bin the incoming
pose to find the correct hash value. Binning is the second
most expensive operation on the fast-path next to calculating
the triangle/box distance.

Speed up the binning by rounding the cayley transform of the
quaternion versor instead of binning by euler angles. It is more
expensive to get the euler angles as it requires several atan2
operations. Getting the cayley transform (and its inverse) requires
negating the versor if the scalar is negative, then a few simple
division/multiplication operations. The rounding error is fairly
uniform across SO(3). For more information, see:

https://marc-b-reynolds.github.io/quaternions/2017/05/02/QuatQuantPart1.html

The old technique exactly matches what the above calls 'ZYX', and
what is implemented is the 'Basic Cayley'. If we ever find that
the cayley transform is introducing too much angular error, another
reasonable comprimise between runtime and accuracy is what the above
calls the 'Basic Harmonic Mean' method.