Navfn: Add params to weight choice of alternate goal
The existing default_tolerance allows Navfn to find the closest
reachable goal to the specified one when it is blocked. However, this
policy can cause it to choose a goal that is on the far side of the
obstacle, possibly requiring a long go-around to reach it. In some
cases, this can lead to long oscillations.

Now the cost of alternate goal is a weighted combination of its
distance to the specified goal and the length of the path required to
get to it (as represented by the potential field value at that point).

The old behavior is obtained by setting tolerance_weight_path_length to
zero (the default).

Note that tolerance_weight_dist_from_goal needs to be much larger than
tolerance_weight_path_length (on the order of 10,000 with a 0.05-meter
map resolution) in order to keep it from always choosing the
shortest-path alternate goal even when the specified goal is clear of
obstacles.