costmap_3d: add nonlethal query, distance options
Add the ability to query nonlethal only obstacles.
A nonlethal obstacle is any Cost between FREE and LETHAL.

Also, extend the distance interface to be passed a DistanceOptions
object. This improves the interface by naming each option.

Add an option to the new interface to limit distance. This is useful
for collision checking or for limiting the distance queries