costmap_3d: maintain query caches across updates
Instead of destroying all the query caches on every update,
maintain the distance and milli-cache to better constrain
subsequent queries on the new map.

The cache mainteance does not bother dealing with new octomap
cells. Instead, it only removes cache entries if their corresponding
octomap cell has been removed. The cache is only to provide an
upper bound to the distance for a query, so this is sufficient.