costmap_3d: add cylinder clearing layer
Add new 3D costmap layer which clears (deletes) any data within a
cylinder around the robot added by previous layers.  This is useful
to implement footprint clearing for global costmaps, which is in turn
useful for global planning in the presence of sensor noise or static
map noise. Because many global planners will operate only on the
resulting 2D map, cylinder clearing is sufficient for many costmap
configurations, and is simple to implement.

The cylinder clearing layer only clears out data from layers above
it, and only by overriding it temporarily. When the robot moves and
uncovers the deleted spot, the old data will be copied back by
forcing it in the bounds map.