costmap_2d: always update master map internal bounds
Always update the stored internal master map bounds on any update.
Previously, the bounds were only saved if there was a change.
This means that if the map did *not* change, the costmap publisher
will mistakenly use the old bounds. This is a big issue on large
maps which change seldomly (such as a static_map only costmap), as
they will publish updates containing the entire map when nothing
changes.