costmap-3d: index tls cache by query obstacle type
The thread local storage distance cache was not being indexed by the
query obstacle type (lethal vs. nonlethal). This is a nasty buy as
the query may incorrectly limit the distance based on an octomap box
that will not be considered. The fix is to index the tls cache by the
obstacle type.

ROS-600 ("Possible Recovery Logic issue 1.18+ (sunlight mitigation
         broken by 3D costmap caching bug)")
https://badgertech.atlassian.net/browse/ROS-600