Revert "costmap_2d: resetBoundingBox: wait until update"
This reverts commit 34551661a30ddeab7d5f978c8cd79a248fca49fe.

That same problem is addressed by this upstream commit:
ec0ef1e7414 ("Add force_updating...param...to clear costmaps recovey")

Furthermore, the change being reverted can cause a ClearCostmapRecovery
to hang indefinitely if we're not receiving LIDAR data becaues the
costmaps remain out of date. If another client tries to initiate a
second ClearCostmapRecovery action when we're running under
MoveBaseFlex, then it will hang trying to cancel the first one and
waiting for it to finish (in pthread.join). At this point, the whole nav
stack will remain permanently stuck even if the LIDAR comes back online
because the single-threaded spinner is unable to process any LIDAR
messages (nor any service calls, etc) because we're stuck in this
pthread.join.

The problem with the change being reverted is that it would not complete
the action until the costmap is current, which can take arbitrary long.
However, the upstream "force_updating" param only waits until we
complete a costmap update. And that's all we actually require. At this
point, the static map has been repopulated into the global costmap and
the 2D inflation is complete so the global planner can behave sanely. If
sensor data is not actually current, we complete the action anyway and a
client may send us a nav goal. But because sensor data is not current,
we won't actually move, so it is OK.

ROS-770
https://badgertech.atlassian.net/browse/ROS-770